refactor: Move code to com.appland.appmap.gradle
