+++ src/main/java/com/appland/appmap/gradle/AppMapPlugin.java
@@ -1,4 +1,4 @@
-package com.appland.appmap;
+package com.appland.appmap.gradle;
 
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
