+++ src/main/java/com/appland/appmap/gradle/AppMapTask.java
@@ -1,4 +1,4 @@
-package com.appland.appmap;
+package com.appland.appmap.gradle;
 
 import org.gradle.api.tasks.testing.Test;
