+++ src/test/java/com/appland/appmap/gradle/AppMapPluginFunctionalTest.java
@@ -1,4 +1,4 @@
-package com.appland.appmap;
+package com.appland.appmap.gradle;
 
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
